Move confirm-email/thank-you binding to upsell.js
