+++ static/src/javascripts/projects/common/modules/identity/upsell/upsell.js
@@ -3,6 +3,7 @@ import { trackNonClickInteraction } from 'common/modules/analytics/google';
 import React, { render } from 'preact-compat';
 import fastdom from 'lib/fastdom-promise';
 import ophan from 'ophan/ng';
+import { ConsentCardList } from 'common/modules/identity/upsell/consent-card/ConsentCardList';
 import loadEnhancers from './../modules/loadEnhancers';
 import { AccountCreationFlow } from './account-creation/AccountCreationFlow';
 import { OptOutsList } from './opt-outs/OptOutsList';
@@ -35,10 +36,20 @@ const bindOptouts = (el): void => {
     });
 };
 
+const bindConfirmEmailThankYou = (el): void => {
+    fastdom.write(() => {
+        render(<ConsentCardList />, el);
+    });
+};
+
 const enhanceUpsell = (): void => {
     loadEnhancers([
         ['.js-identity-upsell-account-creation', bindAccountCreation],
         ['.js-identity-upsell-optputs', bindOptouts],
+        [
+            '.js-identity-upsell-confirm-email-thank-you',
+            bindConfirmEmailThankYou,
+        ],
     ]);
 };
